+++ cmd/curator/main.go
@@ -27,7 +27,7 @@ import (
 )
 
 const (
-	version = "0.24.2"
+	version = "0.25.0"
 )
 
 func main() {